Moved method getNearestWalkableTile(GameObject tileObject) from Rs2WorldPoint to Rs2Tile
